requirements.txt
pillow
django-cleanup